0.4

+++ src/main/java/telegrammrentalbot/rentbot/botAbilities/IBotAbilities.java
@@ -6,7 +6,7 @@ import org.telegram.telegrambots.meta.api.methods.updatingmessages.DeleteMessage
 import java.util.*;
 
 public interface IBotAbilities {
-    DeleteMessage dellMessage(int messageId, long chatId);
+    DeleteMessage dellMessage(long messageId, long chatId);
     void informAll(String text);
     Long sendFullMessageToUser(SendPhoto photo);
     Long sendTextMessageToUserById(String text, long userId);
